Improved compatibility for loading saved shares from older version
