+++ NEWS.md
@@ -24,6 +24,7 @@ Changes
  * Added "Now Playing Sender" plugin for automatically sending names of songs playing to select chat rooms 
  * Builtin plugins load properly on Windows again
  * Modified config backup behavior to not back up the config if "Cancel" is pressed in the file chooser
+ * Shares lists saved in older versions of Nicotine+ can now be loaded again
  * Peer-to-peer (direct) private messaging works properly again
  * General usability improvements to macOS builds
  * Removed option to stop responding to search requests for certain time periods
